+++ lib/src/binary_resources/binary_resources.dart
@@ -1,11 +1,11 @@
 part of binary_marshalling.binary_resources;
 
 class BinaryResources {
-  final Expando _relations = new Expando();
+  static final Expando _relations = new Expando();
 
-  final Expando _resources = new Expando();
+  static final Expando _resources = new Expando();
 
-  BinaryData getData(Object resource) {
+  static BinaryData getData(Object resource) {
     if (resource is int || resource is String || resource is bool || resource == null) {
       throw new ArgumentError.value(resource, "resource");
     }
@@ -13,7 +13,7 @@ class BinaryResources {
     return _resources[resource];
   }
 
-  void registerResource(Object resource, BinaryData data) {
+  static Object registerResource(Object resource, BinaryData data) {
     if (resource is int || resource is String || resource is bool || resource == null) {
       throw new ArgumentError.value(resource, "resource");
     }
@@ -23,9 +23,10 @@ class BinaryResources {
     }
 
     _resources[resource] = data;
+    return resource;
   }
 
-  void registerRelation(Object parent, Object child) {
+  static void registerRelation(Object parent, Object child) {
     if (parent is int || parent is String || parent is bool || parent == null) {
       throw new ArgumentError.value(parent, "parent");
     }
@@ -41,7 +42,7 @@ class BinaryResources {
     _relations[child] = parent;
   }
 
-  void unregisterResource(Object resource) {
+  static void unregisterResource(Object resource) {
     if (resource is int || resource is String || resource is bool || resource == null) {
       throw new ArgumentError.value(resource, "resource");
     }
@@ -50,6 +51,6 @@ class BinaryResources {
       throw new StateError("Resource not found");
     }
 
-    _resources[resource] == null;
+    _resources[resource] = null;
   }
 }